+++ Makefile.am
@@ -1,3 +1,7 @@
 ACLOCAL_AMFLAGS = -I m4
 SUBDIRS = src examples man
 EXTRA_DIST = COPYING autogen.sh
+
+if WANT_EXTRAS
+  SUBDIRS += extras
+endif
