+++ src/Makefile.am
@@ -1,4 +1,4 @@
-# The mux, demux and extras libraries depend on libwebp, thus the '.' to force
+# The mux and demux libraries depend on libwebp, thus the '.' to force
 # the build order so it's available to them.
 SUBDIRS = dec enc dsp utils .
 if WANT_MUX
@@ -6,9 +6,6 @@ if WANT_MUX
 endif
 if WANT_DEMUX
   SUBDIRS += demux
-endif
-if WANT_EXTRAS
-  SUBDIRS += extras
 endif
 
 lib_LTLIBRARIES = libwebp.la
