--- src/extras/libwebpextras.pc.in
@@ -1,11 +1,0 @@
-prefix=@prefix@
-exec_prefix=@exec_prefix@
-libdir=@libdir@
-includedir=@includedir@
-
-Name: libwebpextras
-Description: Additional utility functions useful in processing WebP files
-Version: @PACKAGE_VERSION@
-Requires: libwebp >= 0.5.0
-Cflags: -I${includedir}
-Libs: -L${libdir} -lwebpextras
